Add new tester, does testing lines in order
The tester that is given to us has a flaw where when updating and
selecting lines, it contains the following line of code:

for key in records:

The problem with this is that key could be returned in any order. I.e
the order of the keys returned is not deterministic, so when we run the
output, the keys we get are different than the order the TA had. Thus,
the tester returns a lot of false negatives.

I added a new tester where I copied over the keys exactly as they
appeared in the TA's output, and put them in actual_lines.txt. Thus, the
new_tester.py, instead of the line above, runs the code as the following
pseudocode:

for line in actual_lines.txt:

Thus, when we run this the output does in fact match.