Add Exceptions to functions in Page, fix tests
Added exceptions to most methods in Page for any invalid values or out
of memory errors. This is to help debug any problems later, and to help
point out any bugs we have in the code right now.

Additionally, because of these assertions there were a couple of bugs in
then tests that are now fixed.