Optimize bufferpool, keep track of pid in mem
Added a set to bufferpool that simply tracks which one of the pages we
currently have in memory. This is so when we have to choose one to
vacate, we don't have to loop through all of the page_rep_dir, which was
causing substantial slowdown before.

Readded concurrent merge, tweaked the timeout value.