Change type of lock in bufferpool, remove m3tester
Changed the lock type from a multiprocessing library lock to a
multithreading lock. We do not have bufferpool as a different process,
so this was not needed in the first place.

Additionally, removed m3tester to remove any confusion as to which
tester we should run (it's transaction_tester for now).